+++ mayan/apps/converter/tests/mixins.py
@@ -5,8 +5,10 @@ from mayan.apps.permissions.tests.mixins import PermissionTestMixin
 
 from ..classes import Layer
 from ..models import ObjectLayer
+from ..transformations import BaseTransformation
 
 from .literals import (
+    TEST_TRANSFORMATION_CLASS_LABEL, TEST_TRANSFORMATION_CLASS_NAME,
     TEST_TRANSFORMATION_NAME, TEST_TRANSFORMATION_ARGUMENT,
     TEST_TRANSFORMATION_ARGUMENT_EDITED
 )
@@ -46,7 +48,9 @@ class LayerTestMixin(PermissionTestMixin):
 
 class TransformationTestMixin(LayerTestMixin):
     def _create_test_transformation(self):
-        content_type = ContentType.objects.get_for_model(model=self.test_document)
+        content_type = ContentType.objects.get_for_model(
+            model=self.test_document
+        )
         object_layer, created = ObjectLayer.objects.get_or_create(
             content_type=content_type, object_id=self.test_document.pk,
             stored_layer=self.test_layer.stored_layer
@@ -56,6 +60,23 @@ class TransformationTestMixin(LayerTestMixin):
             name=TEST_TRANSFORMATION_NAME,
             arguments=TEST_TRANSFORMATION_ARGUMENT
         )
+
+    def _create_test_transformation_class(self):
+        class TestTransformation(BaseTransformation):
+            arguments = ('test_argument',)
+            label = TEST_TRANSFORMATION_CLASS_LABEL
+            name = TEST_TRANSFORMATION_CLASS_NAME
+
+            def execute_on(self, *args, **kwargs):
+                super(TestTransformation, self).execute_on(*args, **kwargs)
+
+                return self.image
+
+        BaseTransformation.register(
+            layer=self.test_layer, transformation=TestTransformation
+        )
+
+        self.TestTransformationClass = TestTransformation
 
 
 class TransformationViewsTestMixin(object):
@@ -95,5 +116,25 @@ class TransformationViewsTestMixin(object):
                 'app_label': 'documents', 'model_name': 'document',
                 'object_id': self.test_document.pk,
                 'layer_name': self.test_layer.name
+            }
+        )
+
+    def _request_test_transformation_select_get_view(self):
+        return self.get(
+            viewname='converter:transformation_select', kwargs={
+                'app_label': 'documents', 'model_name': 'document',
+                'object_id': self.test_document.pk,
+                'layer_name': self.test_layer.name
+            }
+        )
+
+    def _request_test_transformation_select_post_view(self):
+        return self.post(
+            viewname='converter:transformation_select', kwargs={
+                'app_label': 'documents', 'model_name': 'document',
+                'object_id': self.test_document.pk,
+                'layer_name': self.test_layer.name
+            }, data={
+                'transformation': TEST_TRANSFORMATION_CLASS_NAME
             }
         )