+++ mayan/apps/converter/tests/test_views.py
@@ -23,7 +23,6 @@ class TransformationViewsTestCase(
         self.grant_access(
             obj=self.test_document, permission=self.test_permission
         )
-
         transformation_count = LayerTransformation.objects.count()
 
         response = self._request_transformation_create_view()
@@ -35,7 +34,6 @@ class TransformationViewsTestCase(
 
     def test_transformation_delete_view_no_permission(self):
         self._create_test_transformation()
-
         transformation_count = LayerTransformation.objects.count()
 
         response = self._request_transformation_delete_view()
@@ -47,11 +45,9 @@ class TransformationViewsTestCase(
 
     def test_transformation_delete_view_with_access(self):
         self._create_test_transformation()
-
         self.grant_access(
             obj=self.test_document, permission=self.test_layer_permission
         )
-
         transformation_count = LayerTransformation.objects.count()
 
         response = self._request_transformation_delete_view()
@@ -63,7 +59,6 @@ class TransformationViewsTestCase(
 
     def test_transformation_edit_view_no_permission(self):
         self._create_test_transformation()
-
         transformation_arguments = self.test_transformation.arguments
 
         response = self._request_transformation_edit_view()
@@ -76,12 +71,11 @@ class TransformationViewsTestCase(
 
     def test_transformation_edit_view_with_access(self):
         self._create_test_transformation()
-
         self.grant_access(
             obj=self.test_document, permission=self.test_layer_permission
         )
-
         transformation_arguments = self.test_transformation.arguments
+
         response = self._request_transformation_edit_view()
         self.assertEqual(response.status_code, 302)
 
@@ -105,7 +99,6 @@ class TransformationViewsTestCase(
 
     def test_transformation_list_view_with_access(self):
         self._create_test_transformation()
-
         self.grant_access(
             obj=self.test_document, permission=self.test_permission
         )
@@ -118,4 +111,54 @@ class TransformationViewsTestCase(
             response=response,
             text=self.test_transformation.get_transformation_class().label,
             status_code=200
+        )
+
+    def test_transformation_select_get_view_no_permission(self):
+        self._create_test_transformation_class()
+        transformation_count = LayerTransformation.objects.count()
+
+        response = self._request_test_transformation_select_get_view()
+        self.assertEqual(response.status_code, 404)
+
+        self.assertEqual(
+            LayerTransformation.objects.count(), transformation_count
+        )
+
+    def test_transformation_select_get_view_with_access(self):
+        self._create_test_transformation_class()
+        self.grant_access(
+            obj=self.test_document, permission=self.test_permission
+        )
+        transformation_count = LayerTransformation.objects.count()
+
+        response = self._request_test_transformation_select_get_view()
+        self.assertEqual(response.status_code, 200)
+
+        self.assertEqual(
+            LayerTransformation.objects.count(), transformation_count
+        )
+
+    def test_transformation_select_post_view_no_permission(self):
+        self._create_test_transformation_class()
+        transformation_count = LayerTransformation.objects.count()
+
+        response = self._request_test_transformation_select_post_view()
+        self.assertEqual(response.status_code, 404)
+
+        self.assertEqual(
+            LayerTransformation.objects.count(), transformation_count
+        )
+
+    def test_transformation_select_post_view_with_access(self):
+        self._create_test_transformation_class()
+        self.grant_access(
+            obj=self.test_document, permission=self.test_permission
+        )
+        transformation_count = LayerTransformation.objects.count()
+
+        response = self._request_test_transformation_select_post_view()
+        self.assertEqual(response.status_code, 302)
+
+        self.assertEqual(
+            LayerTransformation.objects.count(), transformation_count
         )